fix: address PR review (realtime commit race, history load race, mic refresh, crop cast)
- RealtimeTranscriptionService: drain the buffered tail AND claim commitSent in
  one stateQueue critical section, and reject appends once committed/closed, so
  a concurrent appendPCM16 can't strand sub-threshold audio after commit.
- AppState: guard the initial async history load with a mutation-generation
  counter; if a record/retry/delete/clear races the load, re-fetch instead of
  overwriting newer in-memory state. Increment the counter in every mutation.
- AppState: when async mic discovery finishes during recording, set
  needsMicrophoneRefreshAfterRecording instead of dropping the result.
- AppContextService: cast crop-bound origin size to CGFloat explicitly.

+++ Sources/AppContextService.swift
@@ -741,7 +741,12 @@ Selected text: \(selectedText ?? "None")
             y: floor(CGFloat(minY) * scaleY),
             width: ceil(CGFloat(maxX - minX + 1) * scaleX),
             height: ceil(CGFloat(maxY - minY + 1) * scaleY)
-        ).intersection(CGRect(x: 0, y: 0, width: originalWidth, height: originalHeight))
+        ).intersection(CGRect(
+            x: 0,
+            y: 0,
+            width: CGFloat(originalWidth),
+            height: CGFloat(originalHeight)
+        ))
 
         guard !cropRect.isNull, cropRect.width > 0, cropRect.height > 0 else { return image }
